+++ daedalus-server/src/main/java/com/daedalus/server/controller/WorldController.java
@@ -106,7 +106,7 @@ public class WorldController {
                 WorldOps.streetMazes(world), WorldOps.lastLeaseId(world),
                 WorldOps.lastPlaceName(world), WorldOps.occupantsLine(world),
                 WorldOps.standsLine(world), WorldOps.aclLine(world),
-                WorldOps.driveLine(world)));
+                WorldOps.driveLine(world), WorldOps.actorLine(world)));
     }
 
     @GetMapping("/world/{id}/parcels")